Rename CrossSIMD to SIMDHeader, but also keep CrossSIMD.h (will have a future use)
