+++ GPU/Software/Rasterizer.cpp
@@ -36,7 +36,7 @@
 #include "GPU/Software/SoftGpu.h"
 #include "GPU/Software/TransformUnit.h"
 
-#include "Common/Math/CrossSIMD.h"
+#include "Common/Math/SIMDHeaders.h"
 
 namespace Rasterizer {
